build: 26.3-snapshot-7

+++ c2me-base/src/main/java/com/ishland/c2me/base/mixin/access/IBlendingData.java
@@ -36,5 +36,5 @@ public interface IBlendingData {
     HeightLimitView getOldHeightLimit();
 
     @Accessor
-    double[] getSurfaceHeights();
+    float[] getSurfaceHeights();
 }
